test(android): isolate AOSP model path contract
