+++ plugins/plugin-native-inference/__tests__/aosp-model-download-paths.test.ts
@@ -13,7 +13,7 @@ import {
 import {
   bundleSlugFromModelName,
   resolveRecommendedAospModel,
-} from "../src/aosp-local-inference-bootstrap.js";
+} from "../src/aosp-model-paths.js";
 
 describe("AOSP published model resolution", () => {
   it("resolves the first-run chat download through the shared catalog", () => {
