+++ plugins/plugin-native-inference/src/aosp-local-inference-bootstrap.ts
@@ -65,11 +65,7 @@ import {
   type TranscriptionParams,
 } from "@elizaos/core";
 import {
-  buildHuggingFaceResolveUrlForPath,
-  ELIZA_1_TIER_IDS,
-  type Eliza1TierId,
   FIRST_RUN_DEFAULT_MODEL_ID,
-  findCatalogModel,
   tierBundleSlug,
 } from "@elizaos/shared/local-inference";
 import { writeAospLlamaDebugLog } from "./aosp-debug-log.js";
@@ -85,6 +81,10 @@ import {
   createAospStreamingLlmBinding,
   streamGenerate,
 } from "./aosp-llama-streaming.js";
+import {
+  bundleSlugFromModelName,
+  resolveRecommendedAospModel,
+} from "./aosp-model-paths.js";
 import {
   classifyInferenceRamClass,
   InferenceIdleUnloader,
@@ -1300,48 +1300,6 @@ function readBundledModelManifest(modelsDir: string): {
 // `ELIZA_DISABLE_MODEL_AUTO_DOWNLOAD=1` opts out for offline / kiosk
 // builds — callers see the original "stage one via stage-default-models"
 // error in that mode.
-type AospRecommendedModel = {
-  id: string;
-  ggufFile: string;
-  url: string;
-  expectedSizeBytes?: number;
-};
-
-const AOSP_EMBEDDING_TIER_ID = "eliza-1-4b" satisfies Eliza1TierId;
-
-export function resolveRecommendedAospModel(
-  role: "chat" | "embedding",
-): AospRecommendedModel {
-  const tierId =
-    role === "chat" ? FIRST_RUN_DEFAULT_MODEL_ID : AOSP_EMBEDDING_TIER_ID;
-  const model = findCatalogModel(tierId);
-  if (model?.category !== "chat") {
-    throw new Error(
-      `[aosp-local-inference] Catalog is missing ${role} source tier ${tierId}.`,
-    );
-  }
-  if (role === "chat") {
-    const ggufFile = model.hfPathPrefix
-      ? `${model.hfPathPrefix}/${model.ggufFile}`
-      : model.ggufFile;
-    return {
-      id: model.id,
-      ggufFile,
-      url: buildHuggingFaceResolveUrlForPath(model, model.ggufFile),
-    };
-  }
-
-  const ggufFile = `bundles/${tierBundleSlug(tierId)}/embedding/eliza-1-embedding.gguf`;
-  return {
-    id: "eliza-1-embedding",
-    ggufFile,
-    url: buildHuggingFaceResolveUrlForPath(
-      { ...model, hfPathPrefix: undefined },
-      ggufFile,
-    ),
-  };
-}
-
 const aospInflightDownloads = new Map<string, Promise<string>>();
 
 async function downloadRecommendedAospModel(
@@ -1436,27 +1394,6 @@ const KOKORO_VOICE_FILE = "af_sam.bin";
 // and the .bin voice-preset layout). Passed to eliza_inference_kokoro_load.
 const KOKORO_STYLE_DIM = 256;
 let kokoroTtsDownloadInflight: Promise<void> | null = null;
-
-// Derive the current HF bundle tier slug (e.g. "e2b") from a stable chat
-// model id or architecture-slugged GGUF filename. The Kokoro voice URL is
-// `bundles/<tier>/tts/kokoro/...`; the old `path.basename(bundleRoot)`
-// derivation yielded "bundle" for the on-device `<files>/eliza-1/bundle`
-// layout, while the retired size slug (`2b`) no longer names the published
-// Gemma bundle (`e2b`). Defaults to the catalog's first-run bundle slug.
-export function bundleSlugFromModelName(modelNameOrId: string): string {
-  const lower = modelNameOrId.toLowerCase();
-  for (const id of [...ELIZA_1_TIER_IDS].reverse()) {
-    const stableSlug = id.slice("eliza-1-".length);
-    const bundleSlug = tierBundleSlug(id);
-    if (
-      lower.includes(`eliza-1-${stableSlug}`) ||
-      lower.includes(`eliza-1-${bundleSlug}`)
-    ) {
-      return bundleSlug;
-    }
-  }
-  return tierBundleSlug(FIRST_RUN_DEFAULT_MODEL_ID);
-}
 
 // Tier slug of the currently-assigned chat bundle, for the Kokoro voice URL.
 function resolveAssignedChatTierSlug(): string {
